revert consolidating JSDOM stubs
