--- assets/__tests__/helpers.js
@@ -1,26 +1,0 @@
-import { JSDOM } from 'jsdom'
-
-export function liveViewDOM() {
-  const div = document.createElement('div')
-  div.setAttribute('data-phx-view', '')
-  div.setAttribute('data-phx-session', 'abc123')
-  div.setAttribute('id', 'container')
-  div.innerHTML = `
-    <label for="plus">Plus</label>
-    <input id="plus" value="1" />
-    <button phx-click="inc_temperature">Inc Temperature</button>
-  `
-  const button = div.querySelector('button')
-  const input = div.querySelector('input')
-  button.addEventListener('click', () => {
-    setTimeout(() => {
-      input.value += 1
-    }, 200)
-  })
-
-  return div
-}
-
-export function dom() {
-  return new JSDOM(`<!DOCTYPE html><body>${liveViewDOM().outerHTML}</body>`)
-}
